// backend/src/models/payment-response.ts
export interface PaymentResponse {
    PG_TYPE: string;
    addedon: string;
    address1: string;
    address2: string;
    amount: string;
    amount_split: string;
    bank_ref_num: string;
    bankcode: string;
    cardnum: string;
    city: string;
    country: string;
    discount: string;
    email: string;
    encryptedPaymentId: string;
    error: string;
    error_Message: string;
    field1: string;
    field2: string;
    field3: string;
    field4: string;
    field5: string;
    field6: string;
    field7: string;
    field8: string;
    field9: string;
    firstname: string;
    furl: string;
    giftCardIssued: string;
    hash: string;
    isConsentPayment: string;
    key: string;
    lastname: string;
    mihpayid: string;
    mode: string;
    name_on_card: string;
    net_amount_debit: string;
    payuMoneyId: string;
    phone: string;
    productinfo: string;
    state: string;
    status: string;
    surl: string;
    txnMessage: string;
    txnStatus: string;
    txnid: string;
    udf1: string;
    udf2: string;
    udf3: string;
    udf4: string;
    udf5: string;
    udf6: string;
    udf7: string;
    udf8: string;
    udf9: string;
    udf10: string;
    unmappedstatus: string;
    zipcode: string;
}